Move test fixture into __init__
